# Copy-on-write optimization for Clone()