@lemire 2015-09-01
@bpot  Sorry for the delay, merged just now.